e2e status page touchup + refactor util methods to external module
